+++ R/translator.R
@@ -217,34 +217,249 @@ translator_scope_name <- function(x) {
 
 #' Read and Write Translations
 #'
-#' Read [Portable Translator Files] and parse them as [`Translator`][Translator]
-#' objects, or convert [`Translator`][Translator] objects back to
-#' [Portable Translators Files].
+#' @description
+#' Read (import) *Portable Translator Files*, and convert them to
+#' [`Translator`][Translator] objects.
 #'
-#' @template param-path
+#' Write (export) [`Translator`][Translator] objects as
+#' *Portable Translators Files*.
 #'
-#' @template param-encoding
+#' @details
+#' A Portable Translator File (PTF) is a human-friendly, cross language,
+#' and textual representation (serialization) of a [`Translator`][Translator]
+#' object and its underlying translations. The format heavily relies on
+#' [YAML 1.1](https://yaml.org/spec/1.1/), a popular data serialization
+#' format. However, users do not actually have to know anything about YAML
+#' to read, or write [`Translator`][Translator] objects.
+#'
+#' Portable Translator Files are snapshots of [`Translator`][Translator]
+#' objects. To ease collaboration and maintenance, translations are grouped
+#' by language, and stored in separate files. These files are always listed
+#' by the `translations_files` field of a PTF.
+#'
+#' As such, a PTF may refer to a *Portable Translator File*, a
+#' *Portable Translations File*, or both. Portable Translators and Portable
+#' Translations are closely tied by design. It is worthwhile to know that they
+#' both have well-defined internal representations referred to as
+#' [Portable Objects][portable()].
+#'
+#' ## Reading Portable Translations Files
+#'
+#' The easiest way to import translations is to use [translator_read()]. This
+#' function is designed to read all related PTFs, extract their contents, and
+#' construct a suitable [`Translator`][Translator] object from it. This is the
+#' preferred interface. The same is true for [translator_write()].
+#'
+#' Users may read, or write individual Portable Translations Files with
+#' [translations_read()], and [translations_write()], respectively. This
+#' can be useful for debugging purposes.
+#'
+#' ## Encodings
+#'
+#' **The preferred encoding is (and should always be) UTF-8.** If another
+#' encoding must be used for some obscure reason, the `encoding` argument
+#' has to be updated. Encodings should be known, never inferred.
+#'
+#' [translator_write()], and [translations_write()] enforces UTF-8. This
+#' cannot be changed.
+#'
+#' @param path A character string. A path to a file to read from, or write to.
+#'   Its parent directories are automatically created using [dir.create()] if
+#'   they do not exist.
+#'
+#'   By default, [transltor_read()], and [translator_write()] respectively reads
+#'   from, and writes to a default location given by global \R [option][options]
+#'   `transltr.default.path`. It points to a standard
+#'   **Portable Translator File**. See Details below.
+
 #'
 #' @param x An object of class [`Translator`][Translator].
 #'
+#' @template param-encoding
+#'
+#' @template param-lang
+#'
 #' @returns
-#' [translator_import()] returns a [`Translator`][Translator] object if `path`
-#' points to a valid translations source file.
+#' [translator_read()] returns a [`Translator`][Translator] object. This is
+#' an [R6::R6Class()] instance under the hood.
 #'
-#' [translator_export()] is not yet implemented and throws an error.
+#' [translator_write()] returns `NULL`, invisibly. It is used for its
+#' side-effect of creating Portable Translator Files, and Portable Translations
+#' Files.
 #'
-#' @seealso [Portable Translators Files]
+#' [translations_read()] returns an S3 object of class
+#' [`PortableTranslations`][portable()]. Consider using [translator_read()]
+#' instead.
+#'
+#' [translations_write()] returns `NULL`, invisibly. It is used for its
+#' side-effect of creating a single Portable Translations File.
+#'
+#' @seealso
+#' [`Translator`][Translator],
+#' [`Portable`][portable()],
+#' [UTF-8](https://en.wikipedia.org/wiki/UTF-8)
 #'
 #' @rdname translator-io
 #' @export
-translator_import <- function(path = "", encoding = "UTF-8") {
-    return(.NotYetImplemented())
+translator_read <- function(
+    path     = getOption("transltr.default.path"),
+    encoding = "UTF-8")
+{
+    text  <- text_read(path, encoding)
+    trans <- tryCatch({
+        yaml::yaml.load(text,
+            # eval.expr is disallowed for security.
+            eval.expr     = FALSE,
+            as.named.list = TRUE,
+            merge.warning = TRUE,
+            handlers      = list(
+                Translator = as_translator.PortableTranslator,
+                Block      = as_block.PortableBlock,
+                Location   = as_location.PortableLocation))
+    },
+    error = \(cond) {
+        stopf(
+            "while reading 'Translator' object in '%s': %s",
+            path,
+            # Below we attempt to formulate a coherent
+            # sentence from error messages returned by
+            # the YAML parser. This is not perfect, but
+            # better than nothing.
+            gsub("[ \n\t.]+$", ".", tolower(cond$message)))
+    })
+
+    files        <- file.path(dirname(path), attr(trans, "translations_files"))
+    translations <- lapply(files, translations_read, encoding = encoding)
+
+    # Loop over each PortableTranslations, and
+    # for each of them, loop over all non-NULL
+    # translations, and register them.
+    lapply(translations, \(tr) {
+        lang  <- tr$language
+        texts <- lapply(tr$translations, `[[`, i = "translation")
+        texts <- texts[!vapply_1l(texts, is.null)]
+        map(\(hash, text) trans$get_block(hash)$set_translation(lang, text),
+            hash = names(texts),
+            text = texts)
+    })
+
+    return(trans)
 }
 
 #' @rdname translator-io
 #' @export
-translator_export <- function(x, path = "") {
-    return(.NotYetImplemented())
+translator_write <- function(
+    x    = translator(),
+    path = getOption("transltr.default.path"))
+{
+    assert_chr1(path)
+    path       <- normalizePath(path, mustWork = FALSE)
+    parent_dir <- dirname(path)
+
+    if (!dir.exists(parent_dir) &&
+        !dir.create(dirname(parent_dir), TRUE, TRUE) || .__LGL_DEBUG_FLAG) {
+        stopf(
+            "parent directory of path ('%s') could not be created. %s",
+            "Create it manually, or change 'path'.")
+    }
+
+    trans <- portable_translator(x)
+    text_write(format(trans, set_instructions = TRUE), path)
+
+    # Write PortableTranslations files (one
+    # per defined native language). They are
+    # written in the same directory as path.
+    paths <- file.path(dirname(path), trans$translations_files)
+    langs <- names(trans$translations_files)
+    map(translations_write, path = paths, lang = langs, moreArgs = list(x = x))
+
+    return(invisible())
+}
+
+#' @rdname translator-io
+#' @export
+translations_read <- function(path = "", encoding = "UTF-8") {
+    # NOTE: this function complies to the flat format. It
+    # is tightly coupled with format.PortableTranslations().
+
+    text       <- text_read(path, encoding)
+    is_title   <- grepl("^\\[\\[[a-zA-Z0-9]+\\]\\]$", text)
+    sec_starts <- which(is_title)
+    sec_groups <- cumsum(is_title)
+    sec_titles <- gsub("\\[|\\]", "", text[sec_starts])
+
+    # Text is broken into a stream of sections
+    # that can be a YAML header, a source text,
+    # or a translation. Titles are stripped.
+    text <- split_ul(text[-sec_starts], sec_groups[-sec_starts])
+
+    # The first section always corresponds to the
+    # raw YAML header. Other sections are split to
+    # create pairs of source text and translation.
+    header   <- text[[1L]]
+    sections <- split_ul(text[-1L], cumsum(sec_titles != "Translation"))
+    names(sections) <- sec_titles[sec_titles != "Translation"]
+
+    header <- tryCatch({
+        yaml::yaml.load(header,
+            # eval.expr is disallowed for security.
+            eval.expr     = FALSE,
+            as.named.list = TRUE,
+            merge.warning = TRUE)
+    },
+    error = \(cond) {
+        stopf(
+            "while reading YAML header of translations in '%s': %s",
+            path,
+            # Below we attempt to formulate a coherent
+            # sentence from error messages returned by
+            # the YAML parser. This is not perfect, but
+            # better than nothing.
+            gsub("[ \n\t.]+$", ".", tolower(cond$message)))
+    })
+
+    out <- list(
+        language               = header$language,
+        native_language        = header$native_language,
+        source_language        = header$source_language,
+        source_native_language = header$source_native_language,
+        translations           = lapply(sections, \(section) {
+            # There should always be at least one
+            # element that corresponds to source
+            # text by design.
+            source <- section[[1L]]
+            source <- if (!is.null(source)) text_normalize(source)
+
+            # Translation may be missing in the
+            # case of an invalid file format. `[`
+            # returns list(NULL) if this element
+            # does not exist.
+            trans <- section[2L][[1L]]
+            trans <- if (!is.null(trans)) text_normalize(trans)
+
+            return(
+                list(
+                    source_text = source,
+                    translation = if (trans != constant("placeholder")) trans))
+        }))
+
+    return(portable(out, "PortableTranslations", "Translations"))
+}
+
+#' @rdname translator-io
+#' @export
+translations_write <- function(x = translator(), path = "", lang = "") {
+    assert_chr1(path)
+    assert_chr1(lang) # Ensure NULL is always disallowed.
+
+    path    <- normalizePath(path, mustWork = FALSE)
+    strings <- format(
+        # portable_translations() always returns a list.
+        portable_translations(x, lang)[[1L]],
+        how = "flat",
+        set_instructions = TRUE)
+
+    return(text_write(strings, path))
 }